requirements.txt
Pillow==10.0.0
streamlit==1.26.0
